pyproject.toml
[project]
name = "local-mp3-to-spotify"
version = "0.1.0"
description = "Match local MP3/M4A library to Spotify tracks and build playlists."
requires-python = ">=3.11"
dependencies = [
    "spotipy>=2.24.0",
    "mutagen>=1.47.0",
    "python-dotenv>=1.0.1",
    "tqdm>=4.66.0",
    "rapidfuzz>=3.9.0",
]

[tool.uv]
package = false
